Multi tx nonce (#72)
* Multi tx nonce

* Code style
