+++ build.gradle
@@ -1,7 +1,7 @@
 buildscript {
     ext {
         // App version
-        versionName = '1.11.0'
+        versionName = '1.11.1'
         versionCode = 1
 
         // SDK and tools
