+++ fearless-utils/src/test/java/jp/co/soramitsu/fearless_utils/integration/extrinsic/SendIntegrationTest.kt
@@ -11,6 +11,7 @@ import jp.co.soramitsu.fearless_utils.runtime.RealRuntimeProvider
 import jp.co.soramitsu.fearless_utils.runtime.definitions.types.composite.DictEnum
 import jp.co.soramitsu.fearless_utils.runtime.definitions.types.generics.multiAddressFromId
 import jp.co.soramitsu.fearless_utils.runtime.extrinsic.ExtrinsicBuilder
+import jp.co.soramitsu.fearless_utils.runtime.extrinsic.Nonce
 import jp.co.soramitsu.fearless_utils.runtime.extrinsic.signer.KeyPairSigner
 import jp.co.soramitsu.fearless_utils.ss58.SS58Encoder.publicKeyToSubstrateAccountId
 import jp.co.soramitsu.fearless_utils.wsrpc.executeAsync
@@ -39,7 +40,7 @@ class SendIntegrationTest : BaseIntegrationTest(WESTEND_URL) {
                 keypair = KEYPAIR,
                 encryption = MultiChainEncryption.Substrate(EncryptionType.ED25519)
             ),
-            nonce = 38.toBigInteger(),
+            nonce = Nonce.singleTx(38.toBigInteger()),
             runtimeVersion = RuntimeVersion(48, 4),
             genesisHash = "e143f23803ac50e8f6f8e62695d1ce9e4e1d68aa36c1cd2cfd15340213f3423e".fromHex(),
             accountId = KEYPAIR.publicKey.publicKeyToSubstrateAccountId(),
